# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MiniCloud Backend API",
    "version": "1.0.0",
    "description": "Minimal OpenAPI spec for backend with OAuth2 for Swagger UI"
  },
  "servers": [
    { "url": "http://localhost:8081", "description": "Local backend" }
  ],
  "components": {
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "http://localhost:8082/auth/realms/master/protocol/openid-connect/auth",
            "tokenUrl": "http://localhost:8082/auth/realms/master/protocol/openid-connect/token",
            "scopes": {
              "openid": "OpenID scope"
            }
          }
        }
      }
    }
  },
  "security": [ { "oauth2": [ "openid" ] } ],
  "paths": {
    "/hello": {
      "get": {
        "summary": "Hello endpoint",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          }
        }
      }
    },
    "/api/student": {
      "get": {
        "summary": "List students",
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
    "/api/db-test": {
      "get": {
        "summary": "Check database connectivity",
        "responses": {
          "200": {
            "description": "Successfully connected to database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "connected": { "type": "boolean" },
                    "serverTime": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "500": { "description": "Database connection failed" }
        }
      }
    },
    "/api/subjects": {
      "get": {
        "summary": "List subjects from MySQL",
        "responses": {
          "200": {
            "description": "Array of subjects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer" },
                      "name": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "500": { "description": "Database error" }
        }
      }
    },
    "/secure": {
      "get": {
        "summary": "Secure endpoint",
        "responses": {
          "200": { "description": "OK" }
        }
      }
    }
  }
}